Move PathMetadata to path-metadata.ts
