+++ src/lib/reference.ts
@@ -1,13 +1,13 @@
 import type { Blueprint } from '@seamapi/blueprint'
 import type Metalsmith from 'metalsmith'
 
-import type { PathMetadata } from './blueprint.js'
 import {
   type EndpointLayoutContext,
   type RouteLayoutContext,
   setApiRouteLayoutContext,
   setEndpointLayoutContext,
 } from './layout-context.js'
+import { PathMetadataSchema } from './path-metadata.js'
 
 const sdks: Array<'javascript'> = []
 
@@ -19,8 +19,13 @@ export const reference = (
   files: Metalsmith.Files,
   metalsmith: Metalsmith,
 ): void => {
-  const { pathMetadata = {}, ...metadata } =
-    metalsmith.metadata() as Metadata & { pathMetadata: PathMetadata }
+  const metadata = metalsmith.metadata() as Metadata
+
+  // UPSTREAM: Ideally, path metadata would be unnecessary and contained inside the blueprint.
+  const pathMetadata =
+    'pathMetadata' in metadata
+      ? PathMetadataSchema.parse(metadata.pathMetadata)
+      : {}
 
   const blueprint = {
     title: '',
